refactor(utils): rename useQuery to useQueryParams
- Renamed useQuery.ts to useQueryParams.ts
- Updated import paths in FailedView.tsx and related test files
- Adjusted mock implementations to reflect the new hook name
- Ensured all references point to the correct file location
